pyproject.toml
[tool.poetry]
name = "tpmanagement"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13.2"
django = "^5.2.4"
requests = "^2.32.4"
helloasso-apiv5 = "^1.0.0"
dotenv = "^0.9.9"
django-tables2 = "^2.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
